backend/cyber_forensics_service/requirements.txt
fastapi
uvicorn
elasticsearch>=8.0.0
psycopg2-binary
web3
cryptography
python-dotenv